親子関係いけそう、回転と大きさ同時に変更するとおかしくなる

+++ TenkikunEngineProject/MyMath.cpp
@@ -2,10 +2,10 @@
 
 float MyMath::RadToEuler(float rad)
 {
-    return rad * 180.0f / M_PI;
+    return static_cast<float>(rad * 180.0f / M_PI);
 }
 
 float MyMath::EulerToRad(float deg)
 {
-    return deg * M_PI / 180.0f;
+    return static_cast<float>(deg * M_PI / 180.0f);
 }
